+++ js/main.js
@@ -1,4 +1,4 @@
-import './random.js';
+import './utils.js';
 import { getPhotosData } from './users.js';
 import './user-names.js';
 import './comment-length.js';
